convert es6 style for lite ui
